+++ system/src/Doctrine/Common/Cache/FilesystemCache.php
@@ -2,92 +2,23 @@
 
 namespace Doctrine\Common\Cache;
 
+use Grav\Common\Cache\SymfonyCacheProvider;
 use Symfony\Component\Cache\Adapter\FilesystemAdapter;
 
 /**
  * Filesystem cache driver (backwards compatibility).
  */
-class FilesystemCache extends CacheProvider
+class FilesystemCache extends SymfonyCacheProvider
 {
     public const EXTENSION = '.doctrinecache.data';
 
-    /** @var FilesystemAdapter */
-    private $pool;
-
     /**
-     * {@inheritdoc}
+     * @param string $directory
+     * @param string $extension
+     * @param int    $umask
      */
     public function __construct($directory, $extension = self::EXTENSION, $umask = 0002)
     {
-        user_error(self::class . ' is deprecated since Grav 1.8, use Symfony cache instead', E_USER_DEPRECATED);
-
-        $this->pool = new FilesystemAdapter('', 0, $directory);
+        parent::__construct(new FilesystemAdapter('', 0, $directory));
     }
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function doFetch($id)
-    {
-        $item = $this->pool->getItem(rawurlencode($id));
-
-        return $item->isHit() ? $item->get() : false;
-    }
-
-    /**
-     * {@inheritdoc}
-     *
-     * @return bool
-     */
-    protected function doContains($id)
-    {
-        return $this->pool->hasItem(rawurlencode($id));
-    }
-
-    /**
-     * {@inheritdoc}
-     *
-     * @return bool
-     */
-    protected function doSave($id, $data, $lifeTime = 0)
-    {
-        $item = $this->pool->getItem(rawurlencode($id));
-
-        if (0 < $lifeTime) {
-            $item->expiresAfter($lifeTime);
-        }
-
-        return $this->pool->save($item->set($data));
-    }
-
-    /**
-     * {@inheritdoc}
-     *
-     * @return bool
-     */
-    protected function doDelete($id)
-    {
-        return $this->pool->deleteItem(rawurlencode($id));
-    }
-
-    /**
-     * {@inheritdoc}
-     *
-     * @return bool
-     */
-    protected function doFlush()
-    {
-        return $this->pool->clear();
-    }
-
-    /**
-     * {@inheritdoc}
-     *
-     * @return array|null
-     */
-    protected function doGetStats()
-    {
-        return null;
-    }
-
 }